Fixed a bug in the schema CompartmentDefinition
Fixed a bug in the schema CompartmentDefinition. Added the missing UnboundedGemetry element.
Added the SchemaChange description and changed the schemaVersion attribute to v2.8.6_fix

This schema fix is a pre-release schema change and ONLY includes this change. The nexte released schema version from the working_draft branch contains this bug fix.